move numpy_macenko_normalizer to numpy module
